+++ src/main/java/com/mybnb/app/repository/BookingRepository.java
@@ -25,7 +25,7 @@ public interface BookingRepository extends CrudRepository<Booking,Integer>, Book
     @Query(value = "insert into booking (renter_id, listing_id, host_id, start_date, end_date, cost, status) values (?1, ?2, ?3, ?4, ?5, ?6, ?7)", nativeQuery = true)
     @Transactional
     void insertBooking(int renter_id, int listing_id, int host_id, Date start_date,
-        Date end_date, float cost, String status);
+        Date end_date, double cost, String status);
 
     @Query("select b from Booking b where renter_id = ?1")
     List<Booking> fidnByRenterId(int renter_id);
